third_party/blink/web_tests/external/wpt/css/css-anchor-position/anchor-position-multicol-003-expected.txt

This is a testharness.js-based test.
[FAIL] .target 1
  assert_equals: \n<div class="target" data-expected-height="30"></div>\nheight expected 30 but got 0
[FAIL] .target 4
  assert_equals: \n<div class="target" data-expected-height="30"></div>\nheight expected 30 but got 0
[FAIL] .target 5
  assert_equals: \n<div class="target" data-expected-height="30"></div>\nheight expected 30 but got 50
Harness: the test ran to completion.
